Options Menu + String Movement
Howdy Boys; I got the controls and options menu for the start screen added, and I've got the basics of the player control added. They need to be fixed up more, but it's just fiddling with numbers.

If possible, could somebody go and put a max-distance radius that the hand can move from the puppet? It'd really help with the movement!